feat: create fork to version 1.0.4@beta.0 to solve fonts problems
This commit was made to solve fonts problems with latests versions. When font is not found in latests versions, characters are replaced by random
forms. This version uses the disableFontFace props of PDF.js to identify any font of PDF file. It also adds some types to help using this library with
typescript.